requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
motor==3.3.2
pymongo==4.6.3
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyJWT==2.8.0
